minor cosmetic tweaks

+++ src/model/game_id.rs
@@ -48,13 +48,12 @@ impl FromStr for GameId {
 
         let mut n = 0;
         for c in s.bytes().rev() {
-            n = match c {
+            n = u64::from(match c {
                 b'0'..=b'9' => c - b'0',
                 b'A'..=b'Z' => c - b'A' + 10,
                 b'a'..=b'z' => c - b'a' + 10 + 26,
                 _ => return Err(InvalidGameId),
-            } as u64
-                + n * 62;
+            }) + n * 62;
         }
 
         Ok(GameId(n))
